// angular-library/src/app/core/capacity-data/capacity-data.enum.ts
export enum CdmAccordionRoute {
  Ranking = 'rankings',
  WorkingCondition = 'workingconditions',
  WorkingConditionTest = 'workingconditions',
  TapingAndSolidMillingWorkingConditions = 'tapingAndSolidMillingWorkingconditions',
  CorrectionFactor = 'correction-factors',
  CuttingSpeedCorrection = 'cutting-speed-corrections',
  ToolLifeCorrection = 'tool-life-corrections',
  StabilityCorrection = 'stability-corrections',
  CoolantRanking = 'coolant-rankings',
  CoolantCvc = 'coolant-cvc',
  ToolLife = 'tool-life',
  TappingToolLife = 'TappingToolLife',
  DrillingGradeToolLife = 'drillingGradeToolLife',
  CuttingSpeed = 'cuttingspeed',
  StartValues = 'startvalues',
  ReleaseProduct = 'release-product',
  Cd2cValidation = 'cd2c-validation',
  Suboperation = 'suboperations',
  ChipThickness = 'feeds',
  ChipThicknessAxial = 'feeds-axial',
  RelativeEngagement = 'relative-engagement',
  MinSurfaceRoughness = 'surface-roughnesses',
  OperationType = 'operation-types',
  OperationArea = 'operation-area-graphs',
  OperationTypeMedium = 'operationTypeMedium',
  OperationTypeFinishing = 'operationTypeFinishing',
  OperationTypeLowFeed = 'operationTypeLowFeed',
  StartValueKAPR = 'kaprs',
  IndexInsertCount = 'insert-index-counts',
  MaximumOfMaxEngagements = 'maximum-of-max-engagements',
  MaxChipThicknessCorrections = 'max-chip-thickness-corrections',
  AermxCorrections = 'aermx-corrections',
  EngagementCorrections = 'engagement-corrections',
  WorkingEngagementCorrections = 'working-engagement-corrections',
  RampingCorrections = 'ramping-corrections',
  ChipSpaceCorrections = 'chip-space-corrections',
  ToolLifeExponents = 'tool-life-exponents',
  SolidMillingToolLifeCorrection = 'solid-milling-tool-life-corrections',
  CoolantToolLifeCorrection = 'coolant-tool-life-corrections',
  MaximumStress = 'maximum-stress',
  SolidMillingToolLife = 'solid-milling-tool-life',
  SolidMillingCuttingSpeedCorrections = 'solid-milling-cutting-speed-corrections',
  StressCorrections = 'stress-corrections',
  StepOverRatios = 'step-over-ratios',
  RgfRegrindingParameters = 'rgf-regrinding-parameters',
  IndexableMillingSuboperations = 'indexable-milling-supoperation',
  IndexableMillingCuttingSpeedCorrection = 'indexableMillingCuttingSpeedCorrection',
  IndexableMillingToolLifeCorrection = 'indexableMillingToolLifeCorrection',
  IndexableMillingCuttingSpeed = 'IndexableMillingCuttingSpeed',
  FirstChoiceManagement = 'first-choice-management',
  FirstChoice = 'first-choice',
  DiameterFeed = 'diameter-feeds',
  DiameterFeedAxial = 'diameter-feeds-axial',
  FeedFactor = 'feed-factors',
  GeneralTurningBodySuboperations = 'general-turning-body-suboperations',
  GeneralTurningBodyFeedFactor = 'general-turning-body-feed-factors',
  FeedNomFactor = 'feednom-factors',
  Feeds = 'drillingFeeds',
  FeedForceCorrections = 'feed-force-corrections',
  DepthOfCut = 'depths-of-cut',
  CuttingAngle = 'cutting-angles',
  FeedForGeometry = 'feedForGeometry',
  FeedForGrade = 'feedForGrade',
  FeedFroGradeApmx = 'feedForGradeApmx',
  Kapr2Ap = 'kapr-2-aps',
  Kapr3Ap = 'kapr-3-aps',
  Kapr4Ap = 'kapr-4-aps',
  PartingGroovingGeometrySuboperations = 'partingGroovingGeometrySuboperations',
  PartingGroovingGradeSuboperations = 'partingGroovingGradeSuboperations',
  PartingGroovingBodySuboperations = 'partingGroovingBodySuboperations',
  PartingGroovingBodyFeedFactor = 'partingGroovingBodyFeedFactor',
  EnteringAngle = 'entering-angles',
  FeedFlank = 'feeds-flank',
  ThreadTurningBodySuboperations = 'threadTurningBodySuboperations',
  ThreadTurningFeedForGeometry = 'threadTurningFeedForGeometry',
  RecommendedInfeedMethod = 'recommended-infeed-methods',
  OptiThreadRank = 'opti-thread-ranks',
  AdvancedOperation = 'advanced-operations',
  StartValuesFeedFactor = 'feed-label-factors',
  NonRotatingDrilling = 'non-rotating-drillings',
  EntryExitStrategy = 'entry-exit-strategies',
  IndexableDrillingToolLife = 'indexableDrillingToolLife',
  SpeedFactor = 'speed-factors',
  RegrindingParameters = 'regrinding-parameters',
  PeckingPatternStrategy = 'pecking-pattern-strategy',
  DrillingRatio = 'drilling-ratios',
  SolidDrillingGradeToolLife = 'solidDrillingGradeToolLife',
  ProductAssignment = 'product-assignment',
  ChipThicknessCorrection = 'chip-thickness-corrections',
}